grib-template-derive: restructure the source code for clarification
